# Organization/Organization.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organization.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Employee/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "The method returns an employee by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employee"
        ],
        "summary": "The method updates an employee's information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the employee",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the employee",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Employee"
        ],
        "summary": "The method deletes an employee by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the employee",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/EmployeeOccupation": {
      "get": {
        "tags": [
          "EmployeeOccupation"
        ],
        "summary": "The method returns all the connections between Employee and Occupation",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EmployeeOccupation"
        ],
        "summary": "The method adds a new EmployeeOccupation into organization",
        "requestBody": {
          "description": "A new EmployeeOccupation that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeOccupationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeOccupationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeOccupationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/EmployeeOccupation/{id}": {
      "get": {
        "tags": [
          "EmployeeOccupation"
        ],
        "summary": "The method returns a EmployeeOccupation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "EmployeeOccupation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EmployeeOccupation"
        ],
        "summary": "The method updates an EmployeeOccupation information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the EmployeeOccupation",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the EmployeeOccupation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeOccupationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeOccupationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostEmployeeOccupationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EmployeeOccupation"
        ],
        "summary": "The method deletes a EmployeeOccupation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the EmployeeOccupation",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmployeeOccupationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Occupation/{id}": {
      "get": {
        "tags": [
          "Occupation"
        ],
        "summary": "The method returns an occupation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Occupation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOccupationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Occupation"
        ],
        "summary": "The method updates an occupation information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the occupation",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the occupation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostOccupationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostOccupationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostOccupationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOccupationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Occupation"
        ],
        "summary": "The method deletes an occupation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the occupation",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostOccupationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOccupationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOccupationDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetEmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - Identificator for each Employee object",
            "format": "int32"
          },
          "regNumber": {
            "type": "integer",
            "description": "RegNumber - registration number of an Employee",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "description": "FirstName - first name of an Employee",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "LastName - last name of an Employee",
            "nullable": true
          },
          "patronymicName": {
            "type": "string",
            "description": "LastName - patronymic name of on Employee",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "BirthDate - birth date of an Employee",
            "format": "date-time"
          },
          "workshopId": {
            "type": "integer",
            "description": "Workshop - an id of the Workshop",
            "format": "int32"
          },
          "homeAddress": {
            "type": "string",
            "description": "HomeAddress - home address of an employee",
            "nullable": true
          },
          "homeTelephone": {
            "type": "string",
            "description": "HomeTelephone - home telephone of an employee",
            "nullable": true
          },
          "workTelephone": {
            "type": "string",
            "description": "Work Telephone - work telephone of an employee",
            "nullable": true
          },
          "familyStatus": {
            "type": "string",
            "description": "FamilyStatus - family status of an employee (ex. \"married\", \"single\")",
            "nullable": true
          },
          "familyMembersCount": {
            "type": "integer",
            "description": "FamilyMembersCount - number of people in the employee's family",
            "format": "int32"
          },
          "childrenCount": {
            "type": "integer",
            "description": "ChildrenCount - number of the employee's children",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Classs PostEmployeeDto represents an employee in organization, it contains personal information."
      },
      "GetEmployeeOccupationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - an id of the class object",
            "format": "int32"
          },
          "hireDate": {
            "type": "string",
            "description": "HireDate - a date, when an employee was hired on the given occupation.",
            "format": "date-time"
          },
          "dismissalDate": {
            "type": "string",
            "description": "DismissalDate - a date, when an employee was dismissed from the occupation.\r\nCan be null if the employee is still working on that job.",
            "format": "date-time",
            "nullable": true
          },
          "occupationId": {
            "type": "integer",
            "description": "OccupationId - an id of Occupation object",
            "format": "int32",
            "nullable": true
          },
          "employeeId": {
            "type": "integer",
            "description": "EmployeeId - an id of Employee object",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "class PostEmployeeOccupationDto - represents a many-to-many relationship\r\nbetween Employee and Occupation, also it contains date, when an employee was hired, and date, when employee was dismissed."
      },
...
      "GetOccupationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - an id of the occupation",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name - a name of the given occupation",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "PostOccupationDto - represents an employee occupation."
      },
...
      "PostEmployeeDto": {
        "type": "object",
        "properties": {
          "regNumber": {
            "type": "integer",
            "description": "RegNumber - registration number of an Employee",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "description": "FirstName - first name of an Employee",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "LastName - last name of an Employee",
            "nullable": true
          },
          "patronymicName": {
            "type": "string",
            "description": "LastName - patronymic name of on Employee",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "BirthDate - birth date of an Employee",
            "format": "date-time"
          },
          "workshopId": {
            "type": "integer",
            "description": "Workshop - an id of the Workshop",
            "format": "int32"
          },
          "homeAddress": {
            "type": "string",
            "description": "HomeAddress - home address of an employee",
            "nullable": true
          },
          "homeTelephone": {
            "type": "string",
            "description": "HomeTelephone - home telephone of an employee",
            "nullable": true
          },
          "workTelephone": {
            "type": "string",
            "description": "Work Telephone - work telephone of an employee",
            "nullable": true
          },
          "familyStatus": {
            "type": "string",
            "description": "FamilyStatus - family status of an employee (ex. \"married\", \"single\")",
            "nullable": true
          },
          "familyMembersCount": {
            "type": "integer",
            "description": "FamilyMembersCount - number of people in the employee's family",
            "format": "int32"
          },
          "childrenCount": {
            "type": "integer",
            "description": "ChildrenCount - number of the employee's children",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Classs PostEmployeeDto represents an employee in organization, it contains personal information."
      },
      "PostEmployeeOccupationDto": {
        "type": "object",
        "properties": {
          "hireDate": {
            "type": "string",
            "description": "HireDate - a date, when an employee was hired on the given occupation.",
            "format": "date-time"
          },
          "dismissalDate": {
            "type": "string",
            "description": "DismissalDate - a date, when an employee was dismissed from the occupation.\r\nCan be null if the employee is still working on that job.",
            "format": "date-time",
            "nullable": true
          },
          "occupationId": {
            "type": "integer",
            "description": "OccupationId - an id of Occupation object",
            "format": "int32",
            "nullable": true
          },
          "employeeId": {
            "type": "integer",
            "description": "EmployeeId - an id of Employee object",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "class PostEmployeeOccupationDto - represents a many-to-many relationship\r\nbetween Employee and Occupation, also it contains date, when an employee was hired, and date, when employee was dismissed."
      },
...
      "PostOccupationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name - a name of the given occupation",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "PostOccupationDto - represents an employee occupation."
      },
...
    }
  }
}